Refactor `LawExplorer` into modular components
Refactored the `LawExplorer` into smaller, reusable, and modular components such as `LawList`, `LawDetails`, `SearchBar`, and `SimilarLaws`. Introduced utility functions to handle computation logic and moved sample data and types into separate files. Updated dependencies for consistency and added compatibility for enhanced functionality.
