+++ package.json
@@ -10,15 +10,15 @@
     "test": "playwright test"
   },
   "dependencies": {
-    "@radix-ui/react-icons": "1.3.0",
-    "@radix-ui/react-scroll-area": "^1.2.0",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-switch": "^1.1.1",
-    "@radix-ui/react-tabs": "^1.1.1",
-    "class-variance-authority": "^0.7.0",
+    "@radix-ui/react-icons": "1.3.2",
+    "@radix-ui/react-scroll-area": "^1.2.6",
+    "@radix-ui/react-slot": "^1.2.0",
+    "@radix-ui/react-switch": "^1.2.2",
+    "@radix-ui/react-tabs": "^1.1.9",
+    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "framer-motion": "^11.10.0",
-    "lucide-react": "^0.447.0",
+    "framer-motion": "^12.8.0",
+    "lucide-react": "^0.503.0",
     "next": "15.2.4",
     "react": "^19",
     "react-dom": "^19",
@@ -27,6 +27,7 @@
     "tailwindcss-animation": "1.0.1"
   },
   "devDependencies": {
+    "@cloudflare/next-on-pages": "^1.13.12",
     "@playwright/test": "^1.52.0",
     "@types/node": "^20",
     "@types/react": "^19",
